lisatty confirmation alert kun poistetaan kortteja tai kayttajia

+++ views/card.php
@@ -1,3 +1,4 @@
+<script type='text/javascript' src="js/confirmdelete.js"></script>
 <div class="container">
     <h1><?php echo htmlspecialchars($data->card->getName()); ?></h1>
     <?php if ($data->isweapon): ?>
@@ -17,6 +18,6 @@
     <?php endif; ?>
     <?php if (admin()): ?>
     <a href="updatecard.php?id=<?php echo $data->card->getId(); ?>">Edit</a>
-    <a href="destroycard.php?id=<?php echo $data->card->getId(); ?>">Destroy</a>
+    <a href="destroycard.php?id=<?php echo $data->card->getId(); ?>" onclick="return confirmDelete()">Destroy</a>
     <?php endif; ?>
 </div>
