+++ views/user.php
@@ -1,3 +1,4 @@
+<script type='text/javascript' src="js/confirmdelete.js"></script>
 <div class="container">
     <h1><?php echo htmlspecialchars($data->user->getName()); ?></h1>
     <p>Decks made by <?php echo htmlspecialchars($data->user->getName()); ?>:</p>
@@ -8,6 +9,6 @@
     </ul>
     <?php if ($_SESSION["userid"] == $data->user->getId()): ?>
     <a href="updateuser.php?id=<?php echo $data->user->getId(); ?>">Change password</a>
-    <a href="destroyuser.php?id=<?php echo $data->user->getId(); ?>">Destroy</a>
+    <a href="destroyuser.php?id=<?php echo $data->user->getId(); ?>" onclick="return confirmDelete()">Destroy</a>
     <?php endif; ?>
 </div>
